requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.1
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.15.3
